refactor: Remove unused connection testing features
The AWX connection testing functionality has been removed from both the AWX service and the database service. This also includes the removal of the `LogoutIcon` from the Icons component and the removal of the "Connection Testing" feature from the README. The `mock-db-data.json` file has also been cleared as it appears to be leftover test data.