feat: Add Dashboard view to display Ansible facts overview
Introduces a new Dashboard view that provides an aggregated overview of Ansible facts.
This includes:
- Total number of facts.
- Number of unique hosts.
- Distribution of facts by module.
- Top facts by occurrence.

This view enhances the application's ability to quickly understand the landscape of available facts from the AWX instance.
The ChartBarIcon, ServerIcon, and DocumentTextIcon have also been added to the icon library to support this feature.